+++ test/unit/lib/externalUrlUtils.js
@@ -7,107 +7,19 @@ const expect = chai.expect;
 
 describe('externalUrlUtils', () => {
   describe('addMapUrl', () => {
-    const locationDetails = { location: { lat: 52.4, lon: -1.9 } };
+    const location = 'location';
 
-    it('should add additional property to all items in the input list for google maps Url', () => {
-      const nameOne = 'place name one';
-      const nameTwo = 'place name two';
-      const address = {
-        addressLines: ['city', 'county', 'line1', 'line2', 'line3'],
-        postcode: 'AB12 3CD',
-      };
-      const inputList = [{
-        address,
-        name: nameOne,
-      },
-      {
-        address,
-        name: nameTwo,
-      }];
-
-      const startLocation = `saddr=${locationDetails.location.lat}%2C${locationDetails.location.lon}`;
-      const nameAndAddressOne = `${nameOne},${Object.values(address).join()}`;
-      const nameAndAddressOneEncoded = qs.escape(nameAndAddressOne);
-      const destinationOne = `daddr=${nameAndAddressOneEncoded}`;
-      const nearOne = `near=${nameAndAddressOneEncoded}`;
-      const nameAndAddressTwo = `${nameTwo},${Object.values(address).join()}`;
-      const nameAndAddressTwoEncoded = qs.escape(nameAndAddressTwo);
-      const destinationTwo = `daddr=${nameAndAddressTwoEncoded}`;
-      const nearTwo = `near=${nameAndAddressTwoEncoded}`;
-      const encodedQueryOne = `${destinationOne}&${nearOne}&${startLocation}`;
-      const encodedQueryTwo = `${destinationTwo}&${nearTwo}&${startLocation}`;
-      const expectedMapLinkOne = `https://maps.google.com/maps?${encodedQueryOne}`;
-      const expectedMapLinkTwo = `https://maps.google.com/maps?${encodedQueryTwo}`;
-
-      const results = urlUtils.addMapUrl(locationDetails, inputList);
-
-      expect(results).to.not.be.undefined;
-      expect(results).to.be.an('array');
-      expect(results.length).to.be.equal(2);
-      expect(results[0].mapUrl).to.be.equal(expectedMapLinkOne);
-      expect(results[1].mapUrl).to.be.equal(expectedMapLinkTwo);
-    });
-
-    it('should remove any empty, null or undefined address properties', () => {
-      const inputList = [{
-        address: {
-          addressLines: ['', undefined, 'line1', '', null],
-          postcode: 'AB12 3CD',
-        },
-        name: 'name',
-      }];
-
-      const destination = 'name,line1,AB12 3CD';
-      const encodedQuery = `daddr=${qs.escape(destination)}&near=${qs.escape(destination)}&saddr=${locationDetails.location.lat}%2C${locationDetails.location.lon}`;
-      const expectedMapLink = `https://maps.google.com/maps?${encodedQuery}`;
-
-      const results = urlUtils.addMapUrl(locationDetails, inputList);
-
-      expect(results).to.not.be.undefined;
-      expect(results).to.not.be.equal(undefined);
-      expect(results).to.be.an('array');
-      expect(results.length).to.be.equal(1);
-      expect(results[0].mapUrl).to.be.equal(expectedMapLink);
-    });
-
-    it('should not add name as a property if it contains a special symbol', () => {
-      const inputList = [{
-        address: {
-          addressLines: ['', undefined, 'line1', '', null],
-          postcode: 'AB12 3CD',
-        },
-        name: 'name @ subname',
-      }];
-
-      const destination = 'line1,AB12 3CD';
-      const encodedQuery = `daddr=${qs.escape(destination)}&near=${qs.escape(destination)}&saddr=${locationDetails.location.lat}%2C${locationDetails.location.lon}`;
-      const expectedMapLink = `https://maps.google.com/maps?${encodedQuery}`;
-
-      const results = urlUtils.addMapUrl(locationDetails, inputList);
-
-      expect(results).to.not.equal(undefined);
-      expect(results).to.be.an('array');
-      expect(results.length).to.be.equal(1);
-      expect(results[0].mapUrl).to.be.equal(expectedMapLink);
-    });
-
-    it('should use coordinates as start address for location searches', () => {
-      const inputList = [{
-        address: {
-          addressLines: ['line1'],
-          postcode: 'AB12',
-        },
-        name: 'name',
-      }];
-      const destination = 'name,line1,AB12';
+    it('should set \'saddr\' as first param and use the full address from the item for \'daddr\' and \'near\'', () => {
+      const fullAddress = 'line1, AB12';
+      const inputList = [{ address: { fullAddress } }];
       const params = {
-        daddr: destination,
-        near: destination,
-        saddr: `${locationDetails.location.lat},${locationDetails.location.lon}`,
+        daddr: fullAddress,
+        near: fullAddress,
+        saddr: `${location}`,
       };
       const expectedMapLink = `https://maps.google.com/maps?${qs.stringify(params)}`;
 
-      const results = urlUtils.addMapUrl(locationDetails, inputList);
+      const results = urlUtils.addMapUrl(location, inputList);
 
       expect(results.length).to.be.equal(1);
       expect(results[0].mapUrl).to.be.equal(expectedMapLink);
@@ -117,6 +29,7 @@ describe('externalUrlUtils', () => {
   describe('getChoicesResultsUrlToOnlineTests', () => {
     const location = 'ls1';
     const locationDetails = { location: { lat: 52.4, lon: -1.9 } };
+
     it('should return undefined if no location', () => {
       const emptyLocation = '';