+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 - Add pharmacies to kit pickup for 16-24
 - Replace sexperts search with kit locations for over 25s
 - Use `arc` rather than `plane` for precision in distance
+- Use location input for map link
+- Do not use name of place in map link
+- Display address when there is no postcode
 
 0.5.0 / 2018-04-11
 ==================